requirements_basic.txt
azure-functions
requests
beautifulsoup4
validators